rename toggleswitch-with-submit to just toggleswitch since by default it now will not submit unless you give it a prop to
